+++ .gitignore
@@ -11,3 +11,5 @@ node_modules
 .idea
 
 settings.json
+
+.DS_Store
