+++ .meteor/platforms
@@ -1,2 +1,2 @@
-server
 browser
+server
