+++ .meteor/versions
@@ -1,4 +1,5 @@
 accounts-base@1.2.2
+accounts-password@1.1.4
 autoupdate@1.2.4
 babel-compiler@5.8.24_1
 babel-runtime@0.1.4
@@ -23,9 +24,9 @@ diff-sequence@1.0.1
 ecmascript@0.1.6
 ecmascript-runtime@0.2.6
 ejson@1.0.7
+email@1.0.8
 es5-shim@4.1.14
 fastclick@1.0.7
-fourseven:scss@3.4.1
 geojson-utils@1.0.4
 hot-code-push@1.0.0
 html-tools@1.0.5
@@ -40,10 +41,8 @@ localstorage@1.0.5
 logging@1.0.8
 meteor@1.1.10
 meteor-base@1.0.1
-meteorhacks:async@1.0.0
-meteorhacks:fast-render@2.9.0
+meteorhacks:fast-render@2.11.0
 meteorhacks:inject-data@1.4.1
-meteorhacks:npm@1.5.0
 meteorhacks:picker@1.0.3
 minifiers@1.1.7
 minimongo@1.0.10
@@ -51,7 +50,8 @@ mobile-experience@1.0.1
 mobile-status-bar@1.0.6
 mongo@1.1.3
 mongo-id@1.0.1
-npm-container@1.2.0
+nfl:react-helmet@2.2.0
+npm-bcrypt@0.7.8_2
 npm-mongo@1.4.39_1
 observe-sequence@1.0.7
 ordered-dict@1.0.4
@@ -66,19 +66,24 @@ react-runtime-prod@0.14.4
 reactive-dict@1.1.3
 reactive-var@1.0.6
 reactrouter:react-router@1.0.3
-reactrouter:react-router-ssr@1.2.1
+reactrouter:react-router-ssr@1.2.0
 reload@1.1.4
 retry@1.0.4
 routepolicy@1.0.6
-seba:minifiers-autoprefixer@0.0.1
 service-configuration@1.0.5
 session@1.1.1
+sha@1.0.4
 spacebars@1.0.7
 spacebars-compiler@1.0.7
+srp@1.0.4
+standard-minifiers@1.0.2
+thereactivestack:blazetoreact@0.1.5
 tracker@1.0.9
 ui@1.0.8
 underscore@1.0.4
 url@1.0.5
 webapp@1.2.3
 webapp-hashing@1.0.5
-wolves:bourbon@3.1.0
+webpack:npmworkaround@0.1.0
+webpack:reload@0.2.0
+webpack:webpack@0.4.4
