--- client/css/letter-animation.scss
@@ -1,133 +1,0 @@
-@import "{wolves:bourbon}/bourbon";
-
-@import './partials/variables'; // colors, fonts etc...
-
-@import './partials/mixins'; // custom mixins
-
-@import './partials/layout'; // responsive grid and media queries
-
-/* --------------------------------
-
-Primary style
-
--------------------------------- */
-
-.cd-words-wrapper {
-  display: inline-block;
-  position: relative;
-  text-align: left;
-
-  .animated-word {
-    display: inline-block;
-    position: absolute;
-    white-space: nowrap;
-    left: 0;
-    top: 0;
-
-    &.is-visible {
-      position: relative;
-    }
-  }
-
-  .no-js & .animated-word {
-    opacity: 0;
-
-    &.is-visible {
-      opacity: 1;
-    }
-
-  }
-}
-
-
-
-/* --------------------------------
-
-xrotate-2
-
--------------------------------- */
-
-.cd-headline.rotate-2 {
-
-  .cd-words-wrapper {
-    @include perspective(300px);
-  }
-
-  i, em {
-    display: inline-block;
-    @include backface-visibility(hidden);
-  }
-
-  b {
-    opacity: 0;
-  }
-
-  i {
-    @include transform-style(preserve-3d);
-    @include transform(translateZ(-20px) rotateX(90deg));
-    opacity: 0;
-
-    .is-visible & {
-      opacity: 1;
-    }
-
-    &.in {
-      @include animation(cd-rotate-2-in .4s forwards);
-    }
-
-    &.out {
-      @include animation(cd-rotate-2-out .4s forwards);
-    }
-  }
-
-  em {
-    @include transform(translateZ(20px));
-  }
-}
-
-.no-csstransitions .cd-headline.rotate-2 i {
-  @include transform(rotateX(0deg));
-  opacity: 0;
-
-  em {
-    @include transform(scale(1));
-  }
-}
-
-.no-csstransitions .cd-headline.rotate-2 .is-visible i {
-  opacity: 1;
-}
-
-@include keyframes(cd-rotate-2-in) {
-  0% {
-    opacity: 0;
-    @include transform(translateZ(-20px) rotateX(90deg));
-  }
-
-  60% {
-    opacity: 1;
-    @include transform(translateZ(-20px) rotateX(-10deg));
-  }
-
-  100% {
-    opacity: 1;
-    @include transform(translateZ(-20px) rotateX(0deg));
-  }
-}
-
-@include keyframes(cd-rotate-2-out) {
-  0% {
-    opacity: 1;
-    @include transform(translateZ(-20px) rotateX(0));
-  }
-
-  60% {
-    opacity: 0;
-    @include transform(translateZ(-20px) rotateX(-100deg));
-  }
-
-  100% {
-    opacity: 0;
-    @include transform(translateZ(-20px) rotateX(-90deg));
-  }
-}
