--- client/css/partials/_variables.scss
@@ -1,9 +1,0 @@
-// colors
-
-$color-1: #aebcb9; // Tower Gray
-$color-2: #0096a7; // Bondi Blue
-$color-3: #0d0d0d; // Cod Gray
-
-// fonts 
-
-$primary-font: 'Source Sans Pro', sans-serif;
