+++ entry/client/routes.jsx
@@ -1,0 +1,9 @@
+import { Route } from 'react-router';
+
+import Routes from 'App/client/routes'
+
+ReactRouterSSR.Run(
+  <Route>
+    {Routes}
+  </Route>
+);
