+++ entry/webpack.conf.js
@@ -1,0 +1,16 @@
+var webpack = require('webpack');
+var path = require('path');
+
+module.exports = {
+  devServer: {
+    // You can change this to your server IP address to access it remotely
+    host: 'localhost'
+  },
+  hotMiddleware: {
+    reload: true
+  },
+  resolve: {
+    root: path.join(__dirname, '..', 'modules'),
+    extensions: ['', '.js', '.jsx', '.json', '.css', '.scss']
+  }
+};
