--- lib/app.browserify.options.json
@@ -1,11 +1,0 @@
-{
-  "transforms": {
-    "externalify": {
-      "global": true,
-      "external": {
-        "react": "React.require",
-        "react-dom": "React.require"
-      }
-    }
-  }
-}
