+++ modules/App/client/css/login.import.css
@@ -1,0 +1,228 @@
+
+.login-panel {
+    max-width: 360px;
+    background: #fff;
+    box-shadow: 0 -8px 0 -2px rgba(255, 255, 255, 0.7), 0 2px 5px rgba(0, 0, 0, 0.15);
+    border-radius: 8px;
+    color: #666;
+    margin: auto;
+    text-align: center;
+    position: absolute;
+    left: 0;
+    right: 0;
+    top: 0;
+    bottom: 0;
+    height: 450px;
+}
+.login-panel h2 {
+    margin: 25px 0 45px;
+    display: inline-block;
+    text-align: left;
+    border-left: 3px solid;
+    padding-left: 35px;
+    width: 100%;
+    font-weight: 200;
+    color: #0099ed;
+}
+.login-panel a {
+    text-decoration: none;
+    color: #999;
+    display: inline-block;
+    margin-top: 35px;
+}
+.login-panel a:hover {
+    color: #00eda5;
+    text-decoration: underline;
+}
+
+.login-btn {
+    border: 1px solid #ccc;
+    padding: 12px 8px;
+    width: 100%;
+    margin-top: 15px;
+    background: #fff;
+    font-size: 18px;
+    font-weight: 400;
+    cursor: pointer;
+    color: #bbb;
+    margin-bottom: 15px;
+    position: relative;
+    border-radius: 4px;
+}
+.login-btn:hover {
+    color: #fff;
+    border-color: #fff;
+    box-shadow: 0 -100px 0 0 #00eda5 inset;
+}
+
+.register-form {
+    color: #fff;
+    background: #00eda5;
+    font-weight: 100;
+    width: 72px;
+    height: 72px;
+    margin: 5px;
+    padding: 0;
+    border-radius: 100%;
+    overflow: hidden;
+    display: inline-block;
+    cursor: pointer;
+    left: 320px;
+    top: 30px;
+    position: absolute;
+    -webkit-transition: 0.3s;
+    transition: 0.3s;
+}
+.register-form .close {
+    -webkit-transition: 0.3s;
+    transition: 0.3s;
+    position: absolute;
+    top: 8px;
+    width: 50px;
+    height: 50px;
+    z-index: 2;
+    right: 10px;
+    font-size: 50px;
+    -webkit-transform: rotate(45deg);
+    transform: rotate(45deg);
+}
+.register-form .login-formset, .register-form h2 {
+    color: #fff;
+    opacity: 0;
+    position: relative;
+    -webkit-transform: translateX(30px);
+    transform: translateX(30px);
+}
+.register-form h2 {
+    margin-bottom: 15px;
+    -webkit-transform: translateY(30px);
+    transform: translateY(30px);
+}
+.register-form .btn {
+    color: #00eda5;
+    border-color: #fff;
+}
+.register-form .btn:hover {
+    color: #fff;
+}
+.register-form .form-group .form-label, .register-form .form-group .form-control {
+    color: #fff;
+}
+.register-form .form-group.focus:after {
+    background: #fff;
+}
+.register-form .form-group.focus .form-label, .register-form .form-group.filled .form-label {
+    color: #fff;
+}
+.register-form.open {
+    width: 100%;
+    height: 100%;
+    position: absolute;
+    margin: 0;
+    cursor: inherit;
+    z-index: 50;
+    left: 0;
+    top: 0;
+    border-radius: 8px;
+    -moz-transition: width 0.2s 0.2S, height 0.15s 0.25s, border-radius 0.6s, left 0.3s, top 0.3s;
+    -o-transition: width 0.2s 0.2S, height 0.15s 0.25s, border-radius 0.6s, left 0.3s, top 0.3s;
+    -webkit-transition: width 0.2s 0.2S, height 0.15s, border-radius 0.6s, left 0.3s, top 0.3s;
+    -webkit-transition-delay: 0s, 0.25s, 0s, 0s, 0s;
+    -webkit-transition: width 0.2s 0.2S, height 0.15s 0.25s, border-radius 0.6s, left 0.3s, top 0.3s;
+    transition: width 0.2s 0.2S, height 0.15s 0.25s, border-radius 0.6s, left 0.3s, top 0.3s;
+}
+.register-form.open .close {
+    top: 5px;
+    right: 0px;
+    font-size: 32px;
+    opacity: .7;
+    cursor: pointer;
+    z-index: 2;
+    -webkit-transform: rotate(0deg);
+    transform: rotate(0deg);
+}
+.register-form.open .close:hover {
+    opacity: 1;
+}
+.register-form.open .caption {
+    font-size: 50px;
+    margin: 50px auto 15px;
+}
+.register-form.open .login-formset, .register-form.open h2 {
+    opacity: 1;
+    -webkit-transform: translateY(0);
+    transform: translateY(0);
+    -moz-transition: -moz-transform 0.5s 0.55s, opacity 1s 0.55s;
+    -o-transition: -o-transform 0.5s 0.55s, opacity 1s 0.55s;
+    -webkit-transition: -webkit-transform 0.5s, opacity 1s;
+    -webkit-transition-delay: 0.55s, 0.55s;
+    -webkit-transition: opacity 1s 0.55s, -webkit-transform 0.5s 0.55s;
+    transition: opacity 1s 0.55s, -webkit-transform 0.5s 0.55s;
+    transition: transform 0.5s 0.55s, opacity 1s 0.55s;
+    transition: transform 0.5s 0.55s, opacity 1s 0.55s, -webkit-transform 0.5s 0.55s;
+}
+
+.login-formset {
+    padding: 25px;
+    max-width: 320px;
+    margin: auto;
+}
+
+.login-form-group {
+    border-bottom: 1px solid #ddd;
+    position: relative;
+    margin-bottom: 25px;
+}
+.login-form-group:after {
+    content: '';
+    position: absolute;
+    height: 2px;
+    background: #00eda5;
+    width: 0px;
+    left: 0;
+    right: 0;
+    margin: auto;
+    bottom: -2px;
+    opacity: 0;
+    -webkit-transition: 0.3s;
+    transition: 0.3s;
+}
+.login-form-group .login-form-control {
+    border: none;
+    height: 35px;
+    position: relative;
+    z-index: 2;
+    padding: 5px;
+    background: none;
+    color: #00eda5;
+    width: 100%;
+}
+.login-form-group .login-form-control:focus, .login-form-group .login-form-control:active {
+    box-shadow: none;
+    outline: none;
+}
+.login-form-group .login-form-label {
+    font-weight: 300;
+    color: #bbb;
+    font-size: 14px;
+    min-height: 17px;
+    text-align: left;
+    position: absolute;
+    top: 10px;
+    left: 5px;
+    right: 0;
+    -webkit-transition: 0.3s;
+    transition: 0.3s;
+    z-index: 1;
+}
+.login-form-group.focus:after {
+    opacity: 1;
+    width: 100%;
+    background: #00eda5;
+}
+.login-form-group.focus .login-form-label, .login-form-group.filled .login-form-label {
+    top: -7px;
+    font-size: 10px;
+    color: #00eda5;
+}
+
