--- packages.json
@@ -1,5 +1,0 @@
-{
-  "externalify": "0.1.0",
-  "material-ui": "0.14.4",
-  "react-tap-event-plugin": "0.2.2"
-}
