+++ tests/jasmine/client/unit/index.js
@@ -1,0 +1,2 @@
+// Keep this file emoty
+// This file has to exist to run unit tests on client
