+++ tests/jasmine/server/integration/index.js
@@ -1,0 +1,2 @@
+// Keep this file emoty
+// This file has to exist to run integration tests on server
