+++ webpack.packages.json
@@ -1,0 +1,25 @@
+{
+  "babel-core": "^6.3.17",
+  "babel-preset-es2015": "^6.3.13",
+  "babel-preset-stage-0": "^6.3.13",
+  "babel-preset-react": "^6.3.13",
+  "babel-plugin-transform-runtime": "^6.3.13",
+  "babel-plugin-transform-decorators-legacy": "^1.2.0",
+  "babel-plugin-react-transform": "2.0.0-beta1",
+  "babel-runtime": "^6.2.0",
+  "react-mixin": "^3.0.3",
+
+  "babel-loader": "^6.2.0",
+  "null-loader": "^0.1.1",
+  "url-loader": "^0.5.7",
+  "file-loader": "^0.8.5",
+  "style-loader": "^0.13.0",
+  "css-loader": "^0.23.0",
+  "extract-text-webpack-plugin": "^0.9.1",
+
+  "react-transform-hmr": "^1.0.0",
+  "react-transform-catch-errors": "^1.0.0",
+  "redbox-react": "^1.2.0",
+
+  "material-ui": "^0.14.4"
+}
